更改了了显示函数不能为负数的限制

+++ Core/Inc/oled.h
@@ -24,7 +24,7 @@
 void OLED_InitBuffer(void);
 void OLED_ClearBuffer(void);
 void OLED_UpdateDisplayVSync(void); // 更新显示，使用垂直同步
-void OLED_WritePixel(uint8_t x, uint8_t y, uint8_t color); // 写单个像素到缓冲区
+void OLED_WritePixel(int16_t x, int16_t y, uint8_t color); // 写单个像素到缓冲区
 
 // 原始底层函数
 void OLED_SendCommand(uint8_t command);
@@ -33,14 +33,14 @@ void OLED_SendData(uint8_t data);
 void OLED_ClearDisplay(void);
 
 // 绘图函数
-void OLED_SetCursor(uint8_t x, uint8_t y);
-void OLED_DisplayChar(uint8_t x, uint8_t y, char ch);
-void OLED_DisplayString(uint8_t x, uint8_t y, char *str);
-void OLED_DisplayInteger(uint8_t x, uint8_t y, int num);
-void OLED_DisplayFloat(uint8_t x, uint8_t y, float number);
+void OLED_SetCursor(int16_t x, int16_t y);
+void OLED_DisplayChar(int16_t x, int16_t y, char ch);
+void OLED_DisplayString(int16_t x, int16_t y, char *str);
+void OLED_DisplayInteger(int16_t x, int16_t y, int num);
+void OLED_DisplayFloat(int16_t x, int16_t y, float number);
 void OLED_DisplayFPS();
 // 反色文本绘制功能
-void OLED_DisplayCharInverted(uint8_t x, uint8_t y, char ch, uint8_t inverted);
-void OLED_DisplayStringInverted(uint8_t x, uint8_t y, char *str, uint8_t inverted);
+void OLED_DisplayCharInverted(int16_t x, int16_t y, char ch, uint8_t inverted);
+void OLED_DisplayStringInverted(int16_t x, int16_t y, char *str, uint8_t inverted);
 
 #endif /* INC_OLED_H_ */
