+++ Core/Inc/oled_ui.h
@@ -87,11 +87,11 @@ void OLED_MoveObject(AnimationManager_t *manager, const char *tag,
                      uint32_t duration, EaseType_t easeType);
 void OLED_UpdateAnimationManager(AnimationManager_t *manager);
 uint8_t OLED_GetObjectPosition(AnimationManager_t *manager, const char *tag, float *x, float *y);
-uint8_t OLED_GetAnimationStates(AnimationManager_t *manager, const char *tag); // 返回值是1表示活跃，0表示非活跃
+uint8_t OLED_GetAnimationStates(AnimationManager_t *manager, const char *tag);                                                                                              // 返回值是1表示活跃，0表示非活跃
 void OLED_DoTweenObject(AnimationManager_t *manager, const char *tag, float targetX, float targetY, uint32_t duration, EaseType_t easeType, bool enablePrevMutiUseCalling); // 这个函数是用来移动一个对象的，tag是对象的标签，targetX和targetY是目标坐标，duration是动画持续时间，easeType是缓动类型
 void OLED_MoveObjectAxis(AnimationManager_t *manager, const char *tag,
-                          float startX, float startY, float targetX, float targetY,
-                          uint32_t duration, EaseType_t easeType, bool onlyX, bool onlyY); // 为X或Y轴单独创建动画的函数
+                         float startX, float startY, float targetX, float targetY,
+                         uint32_t duration, EaseType_t easeType, bool onlyX, bool onlyY);                                      // 为X或Y轴单独创建动画的函数
 void OLED_DoTweenObjectX(AnimationManager_t *manager, const char *tag, float targetX, uint32_t duration, EaseType_t easeType); // 这个函数是用来移动一个对象的，tag是对象的标签，targetX和targetY是目标坐标，duration是动画持续时间，easeType是缓动类型
 void OLED_DoTweenObjectY(AnimationManager_t *manager, const char *tag, float targetY, uint32_t duration, EaseType_t easeType); // 这个函数是用来移动一个对象的，tag是对象的标签，targetX和targetY是目标坐标，duration是动画持续时间，easeType是缓动类型
 
@@ -102,12 +102,12 @@ void AnimationLoop(void);      // 不使用Manager的动画
 
 // UI绘图函数
 void OLED_DrawTitleBar(char *title);
-void OLED_DrawBattery(uint8_t x, uint8_t y, uint8_t percentage);
-void OLED_DrawProgressBar(uint8_t x, uint8_t y, uint8_t width, uint8_t percent);
+void OLED_DrawBattery(int16_t x, int16_t y, uint8_t percentage);
+void OLED_DrawProgressBar(int16_t x, int16_t y, uint8_t width, uint8_t percent);
 void OLED_DrawRectangle(int16_t x, int16_t y, uint8_t width, uint8_t height);
 void OLED_DrawFilledRectangle(int16_t x, int16_t y, uint8_t width, uint8_t height, uint8_t color);
-void OLED_DrawLine(uint8_t x1, uint8_t y1, uint8_t x2, uint8_t y2);
-void OLED_InvertArea(uint8_t x, uint8_t y, uint8_t width, uint8_t height);
+void OLED_DrawLine(int16_t x1, int16_t y1, int16_t x2, int16_t y2);
+void OLED_InvertArea(int16_t x, int16_t y, uint8_t width, uint8_t height);
 
 #pragma endregion funcs
